lib: trigonometry: move functions as inline in header
These functions are small enough to be inline.
