+++ lib/trigonometry/include/trigonometry.h
@@ -13,8 +13,32 @@
 extern "C" {
 #endif
 
-double limit_angle_rad(double O);
-double limit_angle_deg(double O);
+inline double limit_angle_rad(double O)
+{
+    // TODO: avoid risk of blocking loop
+    while (O > M_PI) {
+        O -= 2.0 * M_PI;
+    }
+
+    while (O < -M_PI) {
+        O += 2.0 * M_PI;
+    }
+
+    return O;
+}
+
+inline double limit_angle_deg(double O)
+{
+    while (O > 180) {
+        O -= 360;
+    }
+
+    while (O < -180) {
+        O += 360;
+    }
+
+    return O;
+}
 
 #ifdef __cplusplus
 }
